refactor: 컴포넌트 분리 및 슬라이드 객체로 관리 - OnBoarding 폴더로 컴포넌트 분리 - OnBoardingList에서 슬라이드 리스트를 객체로 변경
